fix: correct room implementation code

+++ backend/src/modules/room/use-cases/delete-room-use-case.ts
@@ -1,10 +1,10 @@
 import { IDeleteRoomRepository } from "../repository/i-delete-room-respository";
-
+import RoomBusinessException from "@room/infra/exceptions/business-exception";
 
 
 export namespace IDeleteRoomUseCase {
   export type Params = {
-    id: string
+    code: string;
   }
 
   export type Response = {
@@ -26,8 +26,16 @@ export class DeleteRoomsUseCase {
   async execute(props: IDeleteRoomUseCase.Params):
     Promise<IDeleteRoomUseCase.Response> {
 
+    const verifyRoomAlareadyExists = await this.deleteRoomRepository.findByCode({
+      code: props.code,
+    })
+
+    if (!verifyRoomAlareadyExists?.id) {
+      throw new RoomBusinessException("the room does not exist", 404)
+    }
+
     const deleteRoomResponse = await this.deleteRoomRepository.delete({
-      id: props.id,
+      id: verifyRoomAlareadyExists.id!,
     })
 
     return deleteRoomResponse;
